Smaller fixes and touch ups

+++ src/core/intrinsics.h
@@ -26,13 +26,13 @@ inline uint64_t clz(uint64_t v) {
 
 #endif
 
-inline int find_first_zero_byte(uint32_t x) {
+inline uint32_t find_first_zero_byte(uint32_t x) {
     uint32_t y = (x & 0x7F7F7F7F) + 0x7F7F7F7F;
     y = ~(y | x | 0x7F7F7F7F);
     return (31U - clz(y)) >> 3;
 }
 
-inline int find_first_zero_byte(uint64_t x) {
+inline uint64_t find_first_zero_byte(uint64_t x) {
     //uint64_t y = (x & 0x7F7F7F7F7F7F7F7F) + 0x7F7F7F7F7F7F7F7F;
     //y = ~(y | x | 0x7F7F7F7F7F7F7F7F);
     //return (63LLU - clz(y)) >> 3LLU;
